バリデーションのリファクタリング
